Move lib/ to modules/
