+++ modules/getShortcutFromConfig.js
@@ -1,9 +1,7 @@
-function getShortcutFromConfig (key, config) {
+module.exports = function getShortcutFromConfig (key, config) {
   return (
     (config.hotkeys && config.hotkeys[key]) ||
     (config[key] && config[key]['hotkey']) ||
     config[`${key}Shortcut`]
   );
 }
-
-module.exports = { getShortcutFromConfig }
